WIP Add option to upload PDB files
This feel a bit messy. Passing the pdb_string to get_attention_pairs
is only done because if I pass the structure it is hashed and re-calculated
every run.
With newer streamlit this would be easy to fix just add a underscore
infront of the variable name.

It is also confusing now with 3 sources of the pdb data, either
the pdb code, an uploaded file or a pdb string that I store in
session storage to persist uploaded files across page navigation.
I think finding a way to persist the uploaded file would be
easier to understand for the users.