Make layer and head selection less buggy
1. Uses widget state keys instead of explicitly setting session
state keys. Apparently this is more smoot, and I think it seems that way too.
2. Uses a selectbox not number input to encourage selecting a layer in
one operation instead of pressing + or - many times, which would cause
the selected layer to jump back after as the app tried to re-render between
presses.

The re setting of session state at the top of each page is a trick
to get the widget state keys to persist between page navigations.